+++ tests/testthat/test-get_device_info.R
@@ -24,7 +24,7 @@ with_mock_dir("fronius", {
     expect_true(is.data.frame(fronius_info))
     expect_equal(
       names(fronius_info),
-      c("device_id", "last_report", "CustomName", "DT", "ErrorCode", "PVPower", "Show", "StatusCode", "UniqueID")
+      c("device_id", "last_report", "inverter", "CustomName", "DT", "PVPower", "Show", "UniqueID")
       )
     expect_equal(nrow(fronius_info), 1L)
   })
@@ -43,6 +43,30 @@ with_mock_dir("apsystems", {
       c("device_id", "devVer", "ssid", "ipAddr", "minPower", "maxPower")
       )
     expect_equal(nrow(apsystem_info), 2L)
+  })
+})
+
+with_mock_dir("fronius", {
+  test_that("get_device_info() works with multiple devices from Fronius", {
+    skip_on_cran()
+    expect_error(
+      get_device_info(device_ip = c("fronius.local", "fronius2.local"), model = "Fronius"),
+      NA)
+    fronius_info <-  get_device_info(device_ip = c("fronius.local", "fronius2.local"), model = "Fronius")
+    expect_true(is.data.frame(fronius_info))
+    expect_equal(
+      names(fronius_info),
+      c("device_id", "last_report", "inverter", "CustomName", "DT", "PVPower", "Show", "UniqueID")
+      )
+    expect_equal(nrow(fronius_info), 4L)
+  })
+
+  test_that("get_device_info() can raise a warning of one failing out of multiple Fronius", {
+    skip_on_cran()
+    expect_warning(
+      fronius_info <- get_device_info(device_ip = c("fronius.local", "fronius3.local"), model = "Fronius"),
+      "Connection to device")
+    expect_equal(nrow(fronius_info), 1L)
   })
 })